Cleanups, reneabling of the nodebuilder/node for wasm
