--- nodebuilder/node/module_wasm.go
@@ -1,9 +1,0 @@
-//go:build wasm
-
-package node
-
-import "go.uber.org/fx"
-
-func ConstructModule(tp Type) fx.Option {
-	return fx.Module("node")
-}
