+++ nodebuilder/node/node.go
@@ -1,5 +1,3 @@
-//go:build !wasm
-
 package node
 
 import (
